Serializing structure to json being performed via jsonmapper using a map.  Updated templates and js to use correct field names after de-pluralization.  Renamed LoginRequiredController to AccessErrorController.
Added filter which stores users' highest admin rights to session for a limited duration (this supercedes the previous HasViewAdmin filter) and can optionally block non-admins.  Filter is being used in public UI for admin link, admin ui to block non-admins, and services api/status and api/edit.

Added admin error page for users which are logged in but do not have admin rights.  Added solr query method for checking if a user has a specific role granted anywhere in the repository, although it is not being called until we need to be able to differientate between admin types in the admin ui.

Changed api path to /api/, moved all status monitor endpoints to /api/status

Added 401 error page for services